requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
PyPDF2==3.0.1
python-docx==1.1.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
tiktoken==0.5.2
chromadb==0.4.18
pyyaml==6.0.1
pillow==10.1.0
wheel==0.42.0
requests==2.31.0